pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ppmtoconc"
version = "0.1.0"
description = "A Python tool to convert proteomics PPM data into cellular concentrations (g/gDCW)."
authors = [
    { name="Hugues Escoffier", email="hugues.escoffier@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "requests",
    "biopython",
    "bioservices",
    "tqdm"
]

[project.scripts]
ppm-to-concentration = "scripts.run_conversion:main"

[tool.setuptools]
packages = ["ppm_to_conc"]

[project.urls]
"Homepage" = "https://github.com/h-escoffier/PPMtoConc"


